Grass stays standing 16 km from the middle of the world
Reported by the owner: the grass shows for a second or two after a
start and is gone, on WebGL2 and WebGPU alike, flying at 77 m over a
meadow with the panel counting a few thousand tufts. It is not the
ceiling, the GPU or the opening. It is an integer.

A tuft remembers the tile it belongs to by key, tile x times 2^23 plus
tile z, so a tile that leaves the window can take its tufts with it.
The window keeps its set of tiles as numbers, but the per-tuft owners
were an Int32Array, and a key is past 2^31 once the tile is 256 tiles,
16 km, from the middle of the world. There every stored key wrapped,
matched no tile the window wanted, and the first rim rebuild -- the
first 64 m of flight -- threw the whole meadow away, while the window's
own set still said those tiles stood and wrote nothing back. What was
left was the new rim, 400 m out and faded: the few thousand the panel
counted and nobody could see. The owner flies at x -33 789 and -44 584;
a new flight starts near 0, which is why it used to work.

The owners are a Float64Array now, exact far past any tile a flight
reaches. A test flies the window at the owner's own place, rebuilding
four times, and holds it to the meadow a jump there writes: it read
1118 tufts against 9067 before this and reads the same count after.
In a browser at x -44 584 the window holds 22 thousand tufts through
five rebuilds, and they are on the ground.

+++ src/engine/scenery/Grass.ts
@@ -299,8 +299,15 @@ export function createGrass(deps: GrassDeps): Grass {
   });
 
   const standing = new Int32Array(FORMS);
-  /** Which tile each written slot belongs to, so a tile that leaves can take its own tufts with it. */
-  const owner = Array.from({ length: FORMS }, () => new Int32Array(PER_FORM));
+  /**
+   * Which tile each written slot belongs to, so a tile that leaves can take its
+   * own tufts with it. A Float64Array, because a tile's key is not a 32-bit
+   * number: it was an Int32Array, and 16 km out from the middle of the world
+   * the keys wrapped, no standing tuft matched a tile the window wanted, and
+   * the first rim rebuild threw the whole meadow away -- grass for a moment,
+   * then only the new rim, 400 m out and faded.
+   */
+  const owner = Array.from({ length: FORMS }, () => new Float64Array(PER_FORM));
   /** The tiles the window holds, and the ones it wants; the two swap at the end of a rebuild. */
   let live = new Set<number>(),
     wanted = new Set<number>();
@@ -311,8 +318,10 @@ export function createGrass(deps: GrassDeps): Grass {
   /**
    * A tile's name. The window is nineteen tiles a side and the flight covers
    * the world, so the key has to survive tile coordinates in the millions:
-   * multiplying keeps it a small integer where a string would be a small
-   * allocation, once a tile, every rebuild.
+   * multiplying keeps it an integer where a string would be a small
+   * allocation, once a tile, every rebuild. Not a small one: past 256 tiles
+   * out it is beyond 2^31, which is exact in a double and nowhere narrower,
+   * so everything that holds a key holds a double (`owner`).
    */
   const keyOf = (tx: number, tz: number) => tx * 8388608 + tz;
